+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### HEAD
+* Make sure Primary Navigation menu always gets created and has the location set upon activation, update activation permalink method
 * Update to Bootstrap 2.2.1
 * Update conditional feed method
 * Update to Bootstrap 2.2.0